Moved root endpoints to separate folder.
LLMClient class moved to providers/base.py.
