+++ app/main.py
@@ -1,8 +1,6 @@
 from contextlib import asynccontextmanager
 
-from fastapi import FastAPI, status
-from fastapi.responses import RedirectResponse
-from fastapi.templating import Jinja2Templates
+from fastapi import FastAPI
 from fastapi.requests import Request
 from slowapi import _rate_limit_exceeded_handler
 from slowapi.errors import RateLimitExceeded
@@ -12,9 +10,6 @@ from app.rate_limiting import limiter
 from app.config import settings, BASE_PATH
 from app.api import register_routes
 from fastapi.staticfiles import StaticFiles
-
-
-TEMPLATES = Jinja2Templates(directory=str(BASE_PATH / "templates"))
 
 
 @asynccontextmanager
@@ -75,25 +70,6 @@ async def log_requests(request: Request, call_next: callable):
     response = await call_next(request)
     logger.debug(f"Completed with status {response.status_code}")
     return response
-
-
-@app.get("/", status_code=status.HTTP_307_TEMPORARY_REDIRECT)
-@limiter.limit("30/minute")
-async def root(request: Request):
-    return RedirectResponse(url="/ui")
-
-
-@app.get("/ui", status_code=status.HTTP_200_OK)
-@limiter.limit("30/minute")
-async def ui(request: Request):
-    return TEMPLATES.TemplateResponse(
-        "index.html",
-        {
-            "request": request,
-            "minOutTokens": settings.chat.OUTPUT_MIN_TOKENS,
-            "maxOutTokens": settings.chat.OUTPUT_MAX_TOKENS,
-        },
-    )
 
 
 if "__main__" == __name__:
