+++ app/providers/base.py
@@ -5,7 +5,22 @@ must conform to. Using Protocol (structural typing) instead of ABC allows
 for more flexible composition and better type checking support.
 """
 
-from typing import Protocol
+from typing import Protocol, Any
+
+
+class LLMClient(Protocol):
+    """Protocol describing the minimal LLM client surface used by providers.
+
+    Implementations may provide more attributes; this protocol focuses on the
+    members the codebase references (responses, chat, models).
+    """
+
+    responses: Any
+    chat: Any
+    models: Any
+
+    async def aclose(self) -> None:  # optional
+        ...
 
 
 class ToolDefinition(Protocol):
